refactor: changed package name of emailService and otpService to email and otp
